Example with export tariff in
Please note - anything other than last offpeak has to have times (even if full 24h) and if there is start date, it has to be accompanied by end date. In this example I set transitional date on price change, hence it looks silly. Normally it can be used to fluctuate export price based on days or hours, once market comes to it.